+++ src/core/swap/swap-api.ts
@@ -6,6 +6,7 @@ import {
   asMaybeInsufficientFundsError,
   asMaybePendingFundsError,
   asMaybeSwapAboveLimitError,
+  asMaybeSwapAddressError,
   asMaybeSwapBelowLimitError,
   asMaybeSwapCurrencyError,
   asMaybeSwapPermissionError,
@@ -248,6 +249,7 @@ function rankError(error: unknown): number {
   if (asMaybeSwapAboveLimitError(error) != null) return 4
   if (asMaybeSwapPermissionError(error) != null) return 3
   if (asMaybeSwapCurrencyError(error) != null) return 2
+  if (asMaybeSwapAddressError(error) != null) return 2
   return 1
 }