+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: `SwapAddressError` type with `reason` flags
 - changed: Require the React Native environment to include es2015 & `async` support.
 - fixed: `splitWalletInfo` return value when an archived split wallet exists